chore: update processors
- avoid uploading and overwriting existing files on the hub if overwriting is not permitted

+++ vlr/data/processors/uploader.py
@@ -1,5 +1,5 @@
 import os
-from huggingface_hub import HfApi, HfFolder
+from huggingface_hub import HfApi, HfFolder, HfFileSystem
 from vlr.data.utils import zip_dir
 from .processor import Processor
 
@@ -12,12 +12,14 @@ class Uploader(Processor):
         Initialize processor.
         """
         self.api = HfApi()
+        self.fs = HfFileSystem()
 
     def zip_and_upload_dir(
         self, dir_path: str,
         repo_id: str,
         path_in_repo: str,
         repo_type: str = "dataset",
+        overwrite: bool = True,
     ) -> None:
         """
         Zip directory and upload it to the hub.
@@ -26,20 +28,22 @@ class Uploader(Processor):
         :param path_in_repo:    Path to directory in repository.
         :param repo_type:       Repository type.
         """
-        self.api.upload_file(
-            path_or_fileobj=zip_dir(dir_path, overwrite=True),
-            path_in_repo=path_in_repo,
-            repo_id=repo_id,
-            token=self.TOKEN,
-            repo_type=repo_type,
-        )
-        os.remove(dir_path + ".zip")
+        if overwrite or not self.fs.exists(f"{repo_type}s/{repo_id}/{path_in_repo}"):
+            self.api.upload_file(
+                path_or_fileobj=zip_dir(dir_path, overwrite=True),
+                path_in_repo=path_in_repo,
+                repo_id=repo_id,
+                token=self.TOKEN,
+                repo_type=repo_type,
+            )
+            os.remove(dir_path + ".zip")
 
     def upload_file(
         self, file_path: str,
         repo_id: str,
         path_in_repo: str,
         repo_type: str = "dataset",
+        overwrite: bool = True,
     ) -> None:
         """
         Upload file to the hub.
@@ -48,10 +52,11 @@ class Uploader(Processor):
         :param path_in_repo:    Path to file in repository.
         :param repo_type:       Repository type.
         """
-        self.api.upload_file(
-            path_or_fileobj=file_path,
-            path_in_repo=path_in_repo,
-            repo_id=repo_id,
-            token=self.TOKEN,
-            repo_type=repo_type,
-        )
+        if overwrite or not self.fs.exists(f"{repo_type}s/{repo_id}/{path_in_repo}"):
+            self.api.upload_file(
+                path_or_fileobj=file_path,
+                path_in_repo=path_in_repo,
+                repo_id=repo_id,
+                token=self.TOKEN,
+                repo_type=repo_type,
+            )
